move useCurrencies hook to hooks folder
